myfile.requirements.txt
asgiref==3.2.10
Django==3.1.2
django-ckeditor==6.0.0
django-debug-toolbar==3.1.1
django-js-asset==1.2.2
django-mptt==0.11.0
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
